[DITP-pilotage/pilote-2] refactor: package @pilote/kpilote-ui pour composants UI partagés
Crée le package workspace @pilote/kpilote-ui (source-only, calqué sur
kpilote-shared) et y déplace les composants UI génériques jusqu'ici
dupliqués entre kpilote-admin et kpilote-webapp : Button, Typography,
EmptyState, Marianne, Toast, Dialog, Modale, ModaleForm, Picker,
PickerOptionNomId (+ helpers clsxm, searchUnaccent).

Les fichiers ui/ des deux apps deviennent de simples ré-exports du
package : aucun call-site touché, migration à faible risque. Les modales
(webapp) et le Picker (admin) sont désormais disponibles des deux côtés.
Toast/Typography unifiés sur la version webapp (tokens sémantiques
success/accent-rouge, échelle display-xl) — léger changement visuel du
toast de succès admin (bleu → vert). SegmentedControl volontairement
exclu (APIs divergentes entre apps, à réconcilier dans un lot dédié).
Tailwind: directive @source ajoutée pour scanner le package.

Refs:
Project: pilote-2
Client: DITP-pilotage
Category: refactor
Stack: react, typescript, tailwindcss
Tags: design-system, monorepo
Files-changed: 35
