+++ apps/kpilote-admin/package.json
@@ -32,6 +32,7 @@
     "@hono/zod-validator": "^0.7.6",
     "@hookform/resolvers": "^5.4.0",
     "@pilote/kpilote-shared": "workspace:*",
+    "@pilote/kpilote-ui": "workspace:*",
     "@tanstack/react-query": "^5.62.7",
     "@tanstack/react-router": "^1.168.26",
     "@tanstack/react-router-devtools": "^1.166.13",
