+++ packages/kpilote-ui/package.json
@@ -1,0 +1,90 @@
+{
+  "name": "@pilote/kpilote-ui",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./clsxm": {
+      "types": "./src/clsxm.ts",
+      "default": "./src/clsxm.ts"
+    },
+    "./searchUnaccent": {
+      "types": "./src/searchUnaccent.ts",
+      "default": "./src/searchUnaccent.ts"
+    },
+    "./Button": {
+      "types": "./src/Button.tsx",
+      "default": "./src/Button.tsx"
+    },
+    "./Typography": {
+      "types": "./src/Typography.tsx",
+      "default": "./src/Typography.tsx"
+    },
+    "./EmptyState": {
+      "types": "./src/EmptyState.tsx",
+      "default": "./src/EmptyState.tsx"
+    },
+    "./Marianne": {
+      "types": "./src/Marianne.tsx",
+      "default": "./src/Marianne.tsx"
+    },
+    "./Toast": {
+      "types": "./src/Toast.tsx",
+      "default": "./src/Toast.tsx"
+    },
+    "./Dialog": {
+      "types": "./src/Dialog.tsx",
+      "default": "./src/Dialog.tsx"
+    },
+    "./Modale": {
+      "types": "./src/Modale.tsx",
+      "default": "./src/Modale.tsx"
+    },
+    "./ModaleForm": {
+      "types": "./src/ModaleForm.tsx",
+      "default": "./src/ModaleForm.tsx"
+    },
+    "./Picker": {
+      "types": "./src/Picker.tsx",
+      "default": "./src/Picker.tsx"
+    },
+    "./PickerOptionNomId": {
+      "types": "./src/PickerOptionNomId.tsx",
+      "default": "./src/PickerOptionNomId.tsx"
+    }
+  },
+  "files": [
+    "src"
+  ],
+  "scripts": {
+    "lint": "prettier --check .",
+    "format": "prettier --write .",
+    "format:check": "prettier --check ."
+  },
+  "peerDependencies": {
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "^1.1.1",
+    "lucide-react": "^1.14.0",
+    "radix-ui": "^1.4.3",
+    "react": "^19.2.5",
+    "react-dom": "^19.2.5",
+    "react-hook-form": "^7.78.0",
+    "tailwind-merge": "^3.5.0"
+  },
+  "devDependencies": {
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "^1.1.1",
+    "lucide-react": "^1.14.0",
+    "prettier": "^3.6.2",
+    "radix-ui": "^1.4.3",
+    "react": "^19.2.5",
+    "react-dom": "^19.2.5",
+    "react-hook-form": "^7.78.0",
+    "tailwind-merge": "^3.5.0",
+    "typescript": "5.9.3"
+  }
+}
