+++ packages/kpilote-ui/src/clsxm.ts
@@ -1,0 +1,6 @@
+import { clsx, type ClassValue } from 'clsx'
+import { twMerge } from 'tailwind-merge'
+
+export function clsxm(...inputs: ClassValue[]): string {
+  return twMerge(clsx(inputs))
+}
